fix: fix title & icon of BiometricAuth(OK-32960)
